% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumr.R
\name{predict.rumr}
\alias{predict.rumr}
\title{Predict Method for rumr class.}
\usage{
\method{predict}{rumr}(object, predicted, alpha = 0.05, ...)
}
\arguments{
\item{object}{a rumr class object}

\item{predicted}{a numeric vector. Predicted or fitted values from a
a regression model}

\item{alpha}{numeric value. Uncertainty level for the prediction interval.
For instance a 95\% prediction level is given by alpha = 0.05. Confidence
level is defined as 1 - alpha. alpha must be between 0.01 and 0.5.}

\item{...}{...}
}
\description{
Predict Method for rumr class.
}
\author{
David Senhora Navega
}
