% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumr.R
\name{rumr}
\alias{rumr}
\title{Regression Uncertainty Modeler}
\usage{
rumr(
  known,
  predicted,
  type,
  alpha = 0.05,
  interval = NULL,
  delta = 0.25,
  signed = F
)
}
\arguments{
\item{known}{a numeric vector. Known values used to train or fit a
statitical or machine learning regression model}

\item{predicted}{a numeric vector. Predicted values obtain from a regression
model. Ideally this vector should be obtained from cross-validation.}

\item{type}{a character. The type of uncertainty model to fit. See Details}

\item{alpha}{uncertainty level. Used when type = "local".}

\item{interval}{a numeric vector. This vector should have two elements with
the minimum and maximum values with the bounds for the regression function.
Default is NULL, which means that the values are inferred from known vector.
See Details}

\item{delta}{a numeric value. Controls the extrapolation outside of known
values. Must be between 0 and 1. Default is 0.25. See Details.}

\item{signed}{a logical. Defines if conformal prediction should be signed
(assymetrical) or unsigned (symmetrical) prediction intervals around the
predicted value. See Details.}
}
\value{
a rumr class object.
}
\description{
Regression Uncertainty Modeler
}
\details{
TODO
}
\references{
TODO
}
\author{
David Senhora Navega
}
